refac: devide components in license
