+++ src/features/career/constants/license.ts
@@ -1,0 +1,10 @@
+import { License } from "../types/license";
+
+export const licenses: License[] = [
+  {
+    name: "日本商工会議所簿記検定試験2級 合格",
+  },
+  {
+    name: "ITパスポート試験 合格",
+  },
+];
