+++ src/features/career/types/license.ts
@@ -1,0 +1,3 @@
+export type License = {
+  name: string;
+};
